+++ app/decorators/skip_rule_decorator.rb
@@ -1,0 +1,14 @@
+class SkipRuleDecorator < ApplicationDecorator
+  delegate_all
+  # Generates a human readable representation of a skip rule.
+
+  def human_readable(prefs = {})
+    I18n.t("question.skip_rule_instructions", name_and_rank: QuestioningDecorator.new(dest_item).name_and_rank, conditions: decorate_conditions)
+  end
+
+  def decorate_conditions
+    decorated_conditions = ConditionDecorator.decorate_collection(condition_group.members)
+    concatenator = (condition_group.true_if == "all_met") ? I18n.t("common.AND") : I18n.t("common.OR")
+    decorated_conditions.map{ |c| c.human_readable}.join(" #{concatenator} ")
+  end
+end
