+++ app/views/questionings/_printable.html.erb
@@ -28,7 +28,7 @@
 <% qing.skip_rules.each do |rule| %>
   <tr>
     <td colspan="2" class="skip-rule">
-      <%= Printable::SkipRuleDecorator.new(rule).human_readable %>
+      <%= SkipRuleDecorator.new(rule).human_readable %>
     </td>
   </tr>
 <% end %>
